Remove hacky post-format exclusions in favor of homepage template
This removes the hacky post-format exclusion that used to run in
functions.php. The problem with that was that it was hiding
non-standard post formats from XML-RPC calls as well.

This sets a custom homepage template and content query to pull just
standard posts. It was adapted from Twenty Eleven.